Fixed Basename Bug
I don't know how I did this, but the app stopped rendering. I searched through previous commits, trying to find what I did to cause this to happen. Consulted ChatGPT, and it told me to set the basename to the name of my repo a the React router. I did so, and it works again. The only thing I can think of that could have caused it, is something to do with installing gh-pages a short while ago. Investigation ongoing.