Big GSAP Update and Housekeeping
Added an effect that plays a countdown timer anytime a quiz button is pressed. Since quiz buttons are spread out over several components, used a context provider to make the animation available app wide. Added a swipe effect to the quiz so that cards glide out and in based on the button clicked. Cleaned up filenames.